+++ backend/filesystem/anmeldungen.json
@@ -10,6 +10,17 @@
             "street": "null",
             "houseNumber": "null",
             "birthdate": "null"
+        },
+        {
+            "surname": "Sebastian",
+            "lastName": "null",
+            "phoneNumber": "null",
+            "mail": "null",
+            "postCode": "null",
+            "city": "null",
+            "street": "null",
+            "houseNumber": "null",
+            "birthdate": "null"
         }
     ]
 }
